da_experiments/requirements.txt
absl-py==0.11.0
astunparse==1.6.3
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
cox==0.1.post3
cycler==0.10.0
dill==0.3.3
flatbuffers==1.12
gast==0.3.3
gin-config==0.4.0
gitdb==4.0.7
GitPython==3.1.14
google-auth==1.28.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
joblib==1.0.1
jsonpatch==1.28
jsonpointer==2.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.3.2
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
pandas==1.2.3
Pillow==7.2.0
protobuf==3.14.0
psutil==5.8.0
py3nvml==0.2.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.3.1
pyzmq==21.0.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.1
scipy==1.6.0
seaborn==0.11.1
six==1.15.0
smmap==4.0.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.2
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
tfrecord==1.11
threadpoolctl==2.1.0
torch==1.7.1
torchfile==0.1.0
torchvision==0.8.2
tornado==6.0.4
tqdm==4.56.2
typing-extensions==3.7.4.3
urllib3==1.26.3
visdom==0.1.8.9
visdom-logger==0.1
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
xmltodict==0.12.0
